+++ s2-site/examples/layout/custom/demo/custom-table-size.ts
@@ -3,6 +3,7 @@ import { TableSheet } from '@antv/s2';
 fetch('../data/basic.json')
   .then((res) => res.json())
   .then((data) => {
+    // 详情请查看: https://s2.antv.vision/zh/docs/manual/advanced/custom/cell-size
     const container = document.getElementById('container');
     const s2DataConfig = {
       fields: {
@@ -37,6 +38,15 @@ fetch('../data/basic.json')
       width: 600,
       height: 480,
       style: {
+        // 列头宽度始终和数值单元格一致
+        cellCfg: {
+          width: 200,
+          height: 50,
+        },
+        colCfg: {
+          height: 50,
+        },
+        // 每一行根据指标单独设置
         rowCfg: {
           heightByField: {
             '1': 130,
