--- s2-site/examples/layout/custom/demo/custom-width.ts
@@ -1,53 +1,0 @@
-import { PivotSheet } from '@antv/s2';
-
-fetch('../data/basic.json')
-  .then((res) => res.json())
-  .then((data) => {
-    const container = document.getElementById('container');
-    const s2DataConfig = {
-      fields: {
-        rows: ['province', 'city'],
-        columns: ['type'],
-        values: ['price', 'cost'],
-      },
-      meta: [
-        {
-          field: 'province',
-          name: '省份',
-        },
-        {
-          field: 'city',
-          name: '城市',
-        },
-        {
-          field: 'type',
-          name: '商品类别',
-        },
-        {
-          field: 'price',
-          name: '价格',
-        },
-        {
-          field: 'cost',
-          name: '成本',
-        },
-      ],
-      data,
-    };
-
-    const s2Options = {
-      width: 600,
-      height: 480,
-      style: {
-        colCfg: {
-          widthByFieldValue: {
-            price: 90,
-            cost: 50,
-          },
-        },
-      },
-    };
-    const s2 = new PivotSheet(container, s2DataConfig, s2Options);
-
-    s2.render();
-  });
