+++ s2-site/examples/layout/custom/index.zh.md
@@ -1,6 +1,4 @@
 ---
-title: 自定义行高列宽
+title: 自定义行列宽高
 order: 1
 ---
-
-
